+++ src/ingest/auth/refresh/post.ts
@@ -5,12 +5,12 @@
 
 */
 
+import logger from '@frytg/logger'
 import type { Request, Response } from 'express'
 import type { JwtPayload } from 'jsonwebtoken'
 import { DateTime } from 'luxon'
 
 import firebase from '../../../utils/firebase'
-import logger from '../../../utils/logger'
 import response from '../../../utils/response'
 
 const source = 'ingest/auth/refresh'