+++ src/ingest/pubsub/index.ts
@@ -5,9 +5,9 @@
 
 */
 
+import logger from '@frytg/logger'
 import type { Request, Response } from 'express'
 
-import logger from '../../utils/logger'
 import plugins from '../../utils/plugins'
 
 const source = 'ingest/pubsub'