+++ src/ingest/subscriptions/delete.ts
@@ -5,11 +5,11 @@
 
 */
 
+import logger from '@frytg/logger'
 import type { Response } from 'express'
 import type UserTokenRequest from '@/src/ingest/auth/middleware/userTokenRequest.ts'
 
 import datastore from '../../utils/datastore'
-import logger from '../../utils/logger'
 import pubsub from '../../utils/pubsub'
 import response from '../../utils/response'