+++ src/ingest/subscriptions/post.ts
@@ -5,6 +5,7 @@
 
 */
 
+import logger from '@frytg/logger'
 import type { Response } from 'express'
 import { DateTime } from 'luxon'
 import { ulid } from 'ulid'
@@ -13,7 +14,6 @@ import type UserTokenRequest from '@/src/ingest/auth/middleware/userTokenRequest
 import config from '../../../config'
 import { ardFeed } from '../../data/index.ts'
 import datastore from '../../utils/datastore'
-import logger from '../../utils/logger'
 import pubsub from '../../utils/pubsub'
 import response from '../../utils/response'