+++ src/utils/events/processServices.ts
@@ -5,12 +5,12 @@
 
 */
 
+import logger from '@frytg/logger'
 import { createHashedId } from '@swrlab/utils/packages/ard'
 
 import type UserTokenRequest from '@/src/ingest/auth/middleware/userTokenRequest.ts'
 import config from '../../../config'
 import core from '../core'
-import logger from '../logger'
 import pubsub from '../pubsub'
 
 const source = 'utils.events.processServices'