+++ src/utils/firebase/signInWithEmailAndPassword.ts
@@ -5,16 +5,14 @@
 
 */
 
-// load node utils
+import logger from '@frytg/logger'
 import jwt from 'jsonwebtoken'
 
-// load utils
-import logger from '../logger'
 import undici from '../undici'
 
 const source = 'firebase.signInWithEmailAndPassword'
 
-export default async (email:any, password:any) => {
+export default async (email: any, password: any) => {
 	// set firebase sign in url
 	const url = `https://identitytoolkit.googleapis.com/v1/accounts:signInWithPassword?key=${process.env.FIREBASE_API_KEY}`