+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - fix: prevent multiple downloads of the ard feed
 - fix: proper publisher lookup
+- refact: replace logger import
 
 ## [2.0.2] - 2025-10-28